requirements.txt
datasets
faiss-cpu==1.8.0
typer
sentence-transformers==2.6.1
google-generativeai
numpy
nltk
dotenv
pymongo
einops
peft
pymongo 
langchain-community 
llama-index 
llama-index-llms-gemini
llama-index-llms-groq 
llama-index-llms-cohere
langchain-ollama
pandas
llama-index-vector-stores-faiss
llama-index-embeddings-langchain
ollama
aiohttp